Reordering of active memo in DOM
Crude implementation of changing the memos place in the DOM to bring to the top of other memos. This implementation though causes a NotFoundError from functions such as handleBoardDragEnd, this is more than likely caused as the child elements of the board are being wiped and a DocumentFragment being placed in their stead.